+++ dev/npm-publish.sh
@@ -9,7 +9,8 @@ if [[ ! "${VERSION}" =~ ^(patch|minor|major)$ ]]; then
   exit 255
 fi
 
-npm version $VERSION
+npm test
+npm version ${VERSION}
 git push && git push --tag
 npm publish
 ssh 3960.org "cd 3960.org/webcomponent-hsi && git pull && exit"